+++ src/components/common/AuthTooltip/AuthTooltip.test.tsx
