+++ src/components/common/HomeButton/HomeButton.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { useNavigate } from "react-router";
-import home from "../assets/svg/Home.png";
+import home from "../../../assets/svg/Home.png";
 
 const HomeButton = () => {
   const navigate = useNavigate();
