+++ src/pages/FCFSEventPage/FCFSEventPage.tsx
@@ -1,7 +1,14 @@
 import React from "react";
+import { Outlet } from "react-router";
+import HintSection from "./HintSection";
 
 const FCFSEventPage = () => {
-  return <div>FCFSEventPage</div>;
+  return (
+    <div>
+      <Outlet /> {/*QuizSection*/}
+      <HintSection />
+    </div>
+  );
 };
 
 export default FCFSEventPage;
